Update reference arrays with references
Updating reference arrays with references rather than internal models
allows these references to be resolved lazily.

This doesn't have a functional impact for references of a specific type,
because in those cases the default identity map is used and internal
models are built on demand.

However, when using the global cache, reference resolution depends on
the reference already having been populated.  There is no guarantee this
will be the case while a payload is being pushed into the store.  Saving
the reference and resolving it lazily (after invalidating the cp)
ensures that all jsonapi resources that were part of that payload have
made their way into the global cache.

[fix #205]